chore(ui): break opNiceName cycle
