+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/common/opNiceName.ts
@@ -1,0 +1,8 @@
+// Originally was in Links.tsx but that caused a circular dependency
+export const opNiceName = (opName: string) => {
+  let text = opName;
+  if (text.startsWith('op-')) {
+    text = text.slice(3);
+  }
+  return text;
+};
